add: add files for some tasks
